requirements.txt
inference-gpu[yolo-world]==0.9.13
supervision==0.19.0
gradio
tqdm
pillow
ultralytics